requirements.txt
aiogram
asyncpg
python-dotenv

